fix circular dependency issue
